requirements.txt
python-dotenv==1.0.0
pinecone-client==3.0.2
openai==1.12.0
langchain==0.1.4
langchain-openai==0.0.5
PyPDF2==3.0.1 